On quiz query, also return id of creator.
If person is logged in and logged in person's id matches creator id,
show delete button. Unfortunately, delete button doesn't suddenly
disappear after logging off. Delete function turned off if not logged
in.

+++ server/graphql/mutations.js
@@ -54,7 +54,11 @@ const Mutation = types => new GraphQLObjectType({
             args: {
                 id: { type: new GraphQLNonNull(GraphQLID) },
             },
-            resolve(parent, args) {
+            resolve(parent, args, req) {
+                if (!req.user) {
+                    console.log("you need to be logged on to do this"); 
+                    return "You need to be logged on to do this."
+                }
                 Quiz.findByIdAndDelete(args.id)
                     .then(quiz => {
                         console.log(`Deleted quiz ${quiz.title}`);
